requirement.txt
pydantic
beautifulsoup4
requests
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml
streamlit